pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project-ditto-v2"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "datasets>=2.19.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "scipy>=1.13.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    # Terminal-Bench harness (inspect-ai is the Harbor framework dependency)
    "inspect-ai>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
